move get_filtered_balances to ws selectors
