+++ src/components/BalancesTable/BalancesTable.container.js
@@ -1,7 +1,6 @@
 import { connect } from 'react-redux'
 
-import { getAllBalances } from '../../redux/selectors/ws'
-import { getFilteredBalances } from '../../redux/selectors/ui'
+import { getAllBalances, getFilteredBalances } from '../../redux/selectors/ws'
 
 import BalancesTable from './BalancesTable'
