+++ src/redux/selectors/ui/index.js
@@ -16,7 +16,6 @@ import getIsOrderExecuting from './get_is_order_executing'
 import getFirstLogin from './get_first_login'
 import getCurrentMode from './get_current_mode'
 import getGuideStatusForPage from './get_guide_status_for_page'
-import getFilteredBalances from './get_filtered_balances'
 import getStrategyId from './get_strategy_id'
 import getMarketBooks from './get_market_books'
 import getMarketTrades from './get_market_trades'
@@ -40,7 +39,6 @@ export {
   getFirstLogin,
   getCurrentMode,
   getGuideStatusForPage,
-  getFilteredBalances,
   getStrategyId,
   getMarketBooks,
   getMarketTrades,
